update event service

+++ event/src/controllers/event.controller.ts
@@ -1,7 +1,10 @@
 import { EventDetail } from "@event/databases/@types/event.interface";
+import APIError from "@event/Errors/api-error";
 import { validateInput } from "@event/middlewares/input-validation";
+import { verifyToken } from "@event/middlewares/tokenValidation";
 import { EventDetailSchema } from "@event/schemas/event.schema";
 import { EventService } from "@event/services/event.service";
+import { StatusCode } from "@event/utils/consts";
 import {
   Body,
   Controller,
@@ -11,6 +14,7 @@ import {
   Path,
   Post,
   Put,
+  Request,
   Route,
 } from "tsoa";
 
@@ -20,9 +24,14 @@ const eventService = new EventService();
 export class EventController extends Controller {
   @Post("/")
   @Middlewares(validateInput(EventDetailSchema))
-  public async CreateEvent(@Body() requestBody: EventDetail): Promise<any> {
+  @Middlewares(verifyToken)
+  public async CreateEvent(
+    @Body() requestBody: EventDetail,
+    @Request() request: any
+  ): Promise<any> {
     try {
-      const event = await eventService.createEvent(requestBody);
+      const detailEvent = { ...requestBody, orgId: request.id };
+      const event = await eventService.createEvent(detailEvent);
       return {
         message: "Event Created Successfully!",
         data: event,
@@ -33,11 +42,19 @@ export class EventController extends Controller {
   }
 
   @Put("/:id")
+  @Middlewares(verifyToken)
   public async UpdateEvent(
+    @Request() request: any,
     @Path() id: string,
     @Body() requestBody: EventDetail
   ): Promise<any> {
     try {
+      const existedEvent = await eventService.findEventByOrgId(request.id);
+
+      if (!existedEvent) {
+        throw new APIError("Event Not Found !!", StatusCode.NotFound);
+      }
+
       const event = await eventService.updateEvent(id, requestBody);
 
       return {
@@ -49,9 +66,16 @@ export class EventController extends Controller {
     }
   }
 
+  @Middlewares(verifyToken)
   @Delete("/:id")
-  public async DeleteEvent(@Path() id: string) {
+  public async DeleteEvent(@Path() id: string, @Request() request: any) {
     try {
+      const existedEvent = await eventService.findEventByOrgId(request.id);
+
+      if (!existedEvent) {
+        throw new APIError("Event Not Found !!", StatusCode.NotFound);
+      }
+
       await eventService.deleteEvent(id);
       return { message: "Event Deleted Successfully!" };
     } catch (error: unknown) {
@@ -60,14 +84,22 @@ export class EventController extends Controller {
   }
 
   @Get("/:id")
-  public async FindFavoEvent(@Path() id: string): Promise<any> {
+  @Middlewares(verifyToken)
+  public async FindFavoEvent(
+    @Path() id: string,
+    @Request() request: any
+  ): Promise<any> {
     try {
-      const event = await eventService.findEventById(id)
+      const existedEvent = await eventService.findEventByOrgId(request.id);
 
-      return event
+      if (!existedEvent) {
+        throw new APIError("Event Not Found !!", StatusCode.NotFound);
+      }
+      const event = await eventService.findEventById(id);
+
+      return event;
     } catch (error: unknown) {
       throw error;
     }
   }
- 
 }
